pyproject.toml
[project]
name = "mangareadernewgeneration"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.5",
    "asyncpg>=0.30.0",
    "bcrypt>=4.3.0",
    "fastapi[standard]>=0.116.1",
    "minio>=7.2.16",
    "psycopg-binary>=3.2.10",
    "pyjwt>=2.10.1",
    "pytest>=8.4.2",
    "python-decouple>=3.8",
    "python-dotenv>=1.1.1",
    "uvicorn[standard]>=0.35.0",
]
